Update mouse.clj
I'm not sure if I just don't understand how to use this. So I'm uncertain if this args placement I'm proposing is correct. I just can't manage to figure out how to get this function to work and the error that I'm getting in lighttable as I'm trying to use this function is one of apply not having the correct number of arguments.

This is honestly my first time attempting to propose a specific change that's not merely a readme file typo. So I'm fine if I'm just misguided in proposing this.

Thanks